Beginning switch to JxBrowser as the HTML rendering engine
Note that closed-source projects must either (a) use a valid license for
JxBrowser from https://www.teamdev.com/jxbrowser or (b) be modified to
not use JxBrowser. A version without JxBrowser is contained in this
repository for development without JxBrowser, but may not be maintained
as regularly as the version with JxBrowser.